requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.15.1
sqlparse==0.5.5
